clausewise-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pymongo==4.10.1
python-dotenv==1.0.1
bcrypt==3.2.2